Use `num_traits` to add compatibility with `f64`
This commit makes the various structs and functions generic over float
types, and tests that things work for both `f32` and `f64` types.

Unfortunately the code is a little less readable now due to various
trait bounds and generic calls :( There may be a better way of doing
this with `num_traits` that I'm not aware of.