Make StlResults fields public
It's currently hard to store the results of the STL without
using `to_vec()` on the various slices returned by the
seasonal/trend/remainder/weights accessor methods, which
unfortunately requires cloning all the data. By making the
fields public the caller can decide what to do with them.

An alternative could be to add some sort of `into_parts()`
method which returns all four as a tuple or something, but
this just seemed easier!